@@ -2,6 +2,7 @@
   "name": "@storybook/test-runner",
   "version": "0.22.1",
   "description": "Test runner for Storybook stories",
+  "type": "module",
   "keywords": [
     "storybook-addons",
     "test",
@@ -16,11 +17,11 @@
   },
   "license": "MIT",
   "author": "shilman",
-  "main": "dist/index.js",
+  "main": "dist/index.mjs",
   "module": "dist/index.mjs",
   "types": "dist/index.d.ts",
   "bin": {
-    "test-storybook": "./dist/test-storybook.js"
+    "test-storybook": "./dist/test-storybook.mjs"
   },
   "files": [
     "dist",
@@ -55,18 +56,20 @@
     "@babel/generator": "^7.22.5",
     "@babel/template": "^7.22.5",
     "@babel/types": "^7.22.5",
-    "@jest/types": "^29.6.3",
+    "@jest/types": "^30.0.1",
     "@swc/core": "^1.5.22",
-    "@swc/jest": "^0.2.23",
+    "@swc/jest": "^0.2.38",
+    "@types/node-fetch": "^2.6.12",
     "expect-playwright": "^0.8.0",
-    "jest": "^29.6.4",
-    "jest-circus": "^29.6.4",
-    "jest-environment-node": "^29.6.4",
+    "jest": "^30.0.4",
+    "jest-circus": "^30.0.4",
+    "jest-environment-node": "^30.0.4",
     "jest-junit": "^16.0.0",
     "jest-playwright-preset": "^4.0.0",
-    "jest-runner": "^29.6.4",
+    "jest-runner": "^30.0.4",
     "jest-serializer-html": "^7.1.0",
-    "jest-watch-typeahead": "^2.0.0",
+    "jest-watch-typeahead": "^3.0.1",
+    "node-fetch": "^3.3.2",
     "nyc": "^15.1.0",
     "playwright": "^1.14.0"
   },
@@ -80,12 +83,11 @@
     "@storybook/addon-coverage": "^1.0.0",
     "@storybook/addon-docs": "0.0.0-pr-31819-sha-8b752a73",
     "@storybook/react-vite": "0.0.0-pr-31819-sha-8b752a73",
-    "@types/jest": "^29.0.0",
-    "@types/node": "^16.4.1",
-    "@types/node-fetch": "^2.6.11",
+    "@types/jest": "^30.0.0",
+    "@types/node": "^24.0.10",
     "@vitejs/plugin-react": "^4.0.3",
     "auto": "^11.1.6",
-    "babel-jest": "^29.0.0",
+    "babel-jest": "^30.0.4",
     "babel-loader": "^8.1.0",
     "babel-plugin-istanbul": "^6.1.1",
     "can-bind-to-host": "^1.1.1",
@@ -95,7 +97,6 @@
     "husky": "^8.0.0",
     "jest-image-snapshot": "^6.2.0",
     "lint-staged": "^13.0.3",
-    "node-fetch": "^2",
     "pkg-up": "^5.0.0",
     "prettier": "^2.8.1",
     "react": "^17.0.1",
@@ -104,9 +105,9 @@
     "storybook": "0.0.0-pr-31819-sha-8b752a73",
     "tempy": "^1.0.1",
     "ts-dedent": "^2.0.0",
-    "ts-jest": "^29.0.0",
-    "tsup": "^6.5.0",
-    "typescript": "~4.9.4",
+    "ts-jest": "^29.4.0",
+    "tsup": "^8.5.0",
+    "typescript": "^5.8.3",
     "vite": "^6.3.2",
     "wait-on": "^7.2.0"
   },
