@@ -1,4 +1,4 @@
-const { getJestConfig } = require('../dist');
+import { getJestConfig } from '../dist/index.js';
 
 // The default Jest configuration comes from @storybook/test-runner
 const testRunnerConfig = getJestConfig();
@@ -6,7 +6,7 @@ const testRunnerConfig = getJestConfig();
 /**
  * @type {import('@jest/types').Config.InitialOptions}
  */
-module.exports = {
+export default {
   ...testRunnerConfig,
   /** Add your own overrides below, and make sure
    *  to merge testRunnerConfig properties with your own
