@@ -10,23 +10,23 @@ const getTestRunnerPath = () => process.env.STORYBOOK_TEST_RUNNER_PATH ?? '@stor
  * Depending on the user's project and package manager, it's possible that jest-playwright-preset
  * is going to be located in @storybook/test-runner/node_modules OR in the root node_modules
  *
- * By setting `preset: 'jest-playwright-preset` the change of resolution issues is higher, because
+ * By setting `preset: 'jest-playwright-preset' the change of resolution issues is higher, because
  * the lib might be installed inside of @storybook/test-runner/node_modules but references as if it was
  * in the root node_modules.
  *
- * This function does the same thing as `preset: 'jest-playwright-preset` but makes sure that the
+ * This function does the same thing as `preset: 'jest-playwright-preset' but makes sure that the
  * necessary moving parts are all required within the correct path.
  * */
 const getJestPlaywrightConfig = (): Config.InitialOptions => {
   const TEST_RUNNER_PATH = getTestRunnerPath();
   const presetBasePath = path.dirname(
     require.resolve('jest-playwright-preset', {
-      paths: [path.join(__dirname, '../node_modules')],
+      paths: [path.join(import.meta.dirname, '../node_modules')],
     })
   );
   const expectPlaywrightPath = path.dirname(
     require.resolve('expect-playwright', {
-      paths: [path.join(__dirname, '../node_modules')],
+      paths: [path.join(import.meta.dirname, '../node_modules')],
     })
   );
   return {
@@ -54,19 +54,19 @@ export const getJestConfig = (): Config.InitialOptions => {
 
   const jestJunitPath = path.dirname(
     require.resolve('jest-junit', {
-      paths: [path.join(__dirname, '../node_modules')],
+      paths: [path.join(import.meta.dirname, '../node_modules')],
     })
   );
 
   const jestSerializerHtmlPath = path.dirname(
     require.resolve('jest-serializer-html', {
-      paths: [path.join(__dirname, '../node_modules')],
+      paths: [path.join(import.meta.dirname, '../node_modules')],
     })
   );
 
   const swcJestPath = path.dirname(
     require.resolve('@swc/jest', {
-      paths: [path.join(__dirname, '../node_modules')],
+      paths: [path.join(import.meta.dirname, '../node_modules')],
     })
   );
 
@@ -82,11 +82,11 @@ export const getJestConfig = (): Config.InitialOptions => {
     testMatch,
     transform: {
       '^.+\\.(story|stories)\\.[jt]sx?$': require.resolve(
-        `${TEST_RUNNER_PATH}/playwright/transform`
+        `${TEST_RUNNER_PATH}/playwright/transform.js`
       ),
-      '^.+\\.[jt]sx?$': [swcJestPath, { module: { type: 'commonjs' } }],
+      '^.+\\.[jt]sx?$': swcJestPath,
     },
-    transformIgnorePatterns: ['node_modules/(?!(storybook|@storybook)/)'],
+    // transformIgnorePatterns: ['node_modules/(?!(storybook|@storybook)/)'],
     snapshotSerializers: [jestSerializerHtmlPath],
     testEnvironmentOptions: {
       'jest-playwright': {
