@@ -1,14 +1,21 @@
-import { transformSync as swcTransform } from '@swc/core';
+import { transform as swcTransform } from '@swc/core';
 import { transformPlaywright } from './transformPlaywright';
 
-export const process = (src: string, filename: string) => {
-  const csfTest = transformPlaywright(src, filename);
+console.log('FUI CHAMADO');
+throw new Error('teste');
+export const processAsync = async (src: string, filename: string) => {
+  const csfTest = await transformPlaywright(src, filename);
 
-  const result = swcTransform(csfTest, {
+  const result = await swcTransform(csfTest, {
     filename,
+    isModule: true,
     module: {
-      type: 'commonjs',
+      type: 'es6',
     },
   });
   return result ? result.code : src;
+};
+
+export const process = (src: string) => {
+  return src;
 };
